+++ src/Plan.ts
@@ -1,0 +1,7 @@
+type Plan = {
+  id?: number;
+  taskName: string;
+  args: unknown;
+};
+
+export default Plan;