+++ src/index.ts
@@ -1,4 +1,5 @@
 import GuaranteedTask from './GuaranteedTask';
 import TaskRunner from './TaskRunner';
+import TaskChain from './TaskChain';
 
-export { GuaranteedTask, TaskRunner };
+export { GuaranteedTask, TaskRunner, TaskChain };