+++ test/TaskRunner/taskchain.test.ts
@@ -1,34 +1,34 @@
 /* eslint-disable max-classes-per-file */
-import { TaskRunner } from '../../src/index';
+import { TaskRunner, TaskChain } from '../../src/index';
 import { waitForNextLoop } from '../utils';
 import CreateNormalTask from '../NormalTask';
 import CreateChainTask from '../ChainTask';
 
-const mock = jest.fn();
-const NormalTask = CreateNormalTask(mock);
-const ChainTask = CreateChainTask(mock);
+describe('task chain normal', () => {
+  const mock = jest.fn();
+  const NormalTask = CreateNormalTask(mock);
+  const ChainTask = CreateChainTask(mock);
 
-let taskRunner: TaskRunner;
-beforeEach(() => {
-  taskRunner = new TaskRunner({
-    Tasks: [ChainTask, NormalTask],
-    dbOptions: { name: ':memory:' },
+  let taskRunner: TaskRunner;
+  beforeEach(() => {
+    taskRunner = new TaskRunner({
+      Tasks: [ChainTask, NormalTask],
+      dbOptions: { name: ':memory:' },
+    });
   });
-});
-afterEach(async () => {
-  await taskRunner.stop();
-  taskRunner.db.close();
-});
+  afterEach(async () => {
+    await taskRunner.stop();
+    taskRunner.db.close();
+  });
 
-describe('task chaining', () => {
   it('Runs all tasks of the chain', async () => {
     await taskRunner.start();
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    taskRunner
-      .add(NormalTask, 'first')
-      .then(NormalTask, 'second')
-      .then(NormalTask, 'third')
-      .exec();
+    taskRunner.execute(
+      new TaskChain(NormalTask, 'first')
+        .add(NormalTask, 'second')
+        .add(NormalTask, 'third'),
+    );
     await waitForNextLoop(5);
     await taskRunner.stop();
     expect(mock).toBeCalledTimes(3);
@@ -40,7 +40,9 @@ describe('task chaining', () => {
   it('Passes args to the next', async () => {
     await taskRunner.start();
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    taskRunner.add(ChainTask, 1).then(ChainTask).then(ChainTask).exec();
+    taskRunner.execute(
+      new TaskChain(ChainTask, 1).add(ChainTask).add(ChainTask),
+    );
     await waitForNextLoop(5);
     await taskRunner.stop();
     expect(mock).toBeCalledTimes(3);
@@ -52,7 +54,9 @@ describe('task chaining', () => {
   it("doesn't pass args to the next if next task has args set", async () => {
     await taskRunner.start();
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    taskRunner.add(ChainTask, 1).then(ChainTask).then(ChainTask, 5).exec();
+    taskRunner.execute(
+      new TaskChain(ChainTask, 1).add(ChainTask).add(ChainTask, 5),
+    );
     await waitForNextLoop(5);
     await taskRunner.stop();
     expect(mock).toBeCalledTimes(3);
@@ -60,11 +64,53 @@ describe('task chaining', () => {
     expect(mock).toBeCalledWith(2);
     expect(mock).toBeCalledWith(5);
   });
+});
+
+describe('task chain debug', () => {
+  const mock = jest.fn();
+  const NormalTask = CreateNormalTask(mock);
+  const ChainTask = CreateChainTask(mock);
+
+  let taskRunner: TaskRunner;
+  beforeEach(() => {
+    taskRunner = new TaskRunner({
+      Tasks: [ChainTask, NormalTask],
+      dbOptions: { name: ':memory:' },
+      debug: {
+        runLimit: 1,
+      },
+    });
+  });
+  afterEach(async () => {
+    await taskRunner.stop();
+    taskRunner.db.close();
+  });
+
+  it("doesn't run until started", async () => {
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    taskRunner.execute(
+      new TaskChain(ChainTask, 1).add(ChainTask).add(ChainTask),
+    );
+    await taskRunner.start();
+    await taskRunner.stop();
+    expect(mock).toBeCalledTimes(1);
+    expect(mock).toBeCalledWith(1);
+    await taskRunner.start();
+    await taskRunner.stop();
+    expect(mock).toBeCalledTimes(2);
+    expect(mock).toBeCalledWith(2);
+    await taskRunner.start();
+    await taskRunner.stop();
+    expect(mock).toBeCalledTimes(3);
+    expect(mock).toBeCalledWith(3);
+  });
 
   it('continues chain after restart', async () => {
     await taskRunner.start();
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    taskRunner.add(ChainTask, 1).then(ChainTask).then(ChainTask).exec();
+    taskRunner.execute(
+      new TaskChain(ChainTask, 1).add(ChainTask).add(ChainTask),
+    );
     await taskRunner.stop();
     expect(mock).toBeCalledTimes(1);
     expect(mock).toBeCalledWith(1);
@@ -81,29 +127,34 @@ describe('task chaining', () => {
   it('continues chain properly with interruptions', async () => {
     await taskRunner.start();
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    taskRunner.add(ChainTask, 1).then(ChainTask).then(ChainTask).exec();
+    taskRunner.execute(
+      new TaskChain(ChainTask, 1).add(ChainTask).add(ChainTask),
+    );
     await taskRunner.stop();
-    await waitForNextLoop(100);
     expect(mock).toBeCalledTimes(1);
     expect(mock).toBeCalledWith(1);
     await taskRunner.start();
-    await waitForNextLoop(100);
+    await taskRunner.stop();
+    expect(mock).toBeCalledTimes(2);
+    expect(mock).toBeCalledWith(2);
+    await taskRunner.start();
     await taskRunner.stop();
     expect(mock).toBeCalledTimes(3);
     expect(mock).toBeCalledWith(3);
   });
 
   it('removes rest of the chain if a failed task removes itself', async () => {
-    await taskRunner.start();
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    taskRunner
-      .add(ChainTask, 1) // 1
-      .then(ChainTask) // 2
-      .then(ChainTask) // 3
-      .then(ChainTask) // 4 and fail on this one
-      .then(ChainTask) // 5
-      .then(ChainTask) // 6
-      .exec();
+    await taskRunner.execute(
+      new TaskChain(ChainTask, 1) // 1
+        .add(ChainTask) // 2
+        .add(ChainTask) // 3
+        .add(ChainTask) // 4 and fail on this one
+        .add(ChainTask) // 5
+        .add(ChainTask), // 6
+    );
+    expect(mock).toBeCalledTimes(0);
+    await taskRunner.start();
     await taskRunner.stop();
     expect(mock).toBeCalledTimes(1);
     expect(mock).toBeCalledWith(1);